+++ example/upsert_test.dart
@@ -2,22 +2,23 @@ import 'package:litesql/litesql.dart';
 import 'package:println/println.dart';
 import 'package:sqlite3/sqlite3.dart';
 
-void main() async {
+void main2() async {
   LiteSQL lite = LiteSQL.openMemory();
   lite.lastInsertRowId = 8;
   // clearLastRowId(lite.database.handle, 2);
   println("last insert row: ", lite.lastInsertRowId);
 }
 
-void main2() async {
+void main() async {
   LiteSQL lite = LiteSQL.openMemory();
+  println("version: ", LiteSQL.version.versionNumber);
   lite.executeSQL("CREATE TABLE stu(id INTEGER PRIMARY KEY AUTOINCREMENT, name text)");
 
   ResultSet r1 = lite.selectSQL("INSERT INTO stu(name) values('yang')  RETURNING id");
   println("lastRowID: ", lite.lastInsertRowId);
   r1.dump();
 
-  lite.setLastInsertRowId(0);
+  lite.lastInsertRowId = 0;
   ResultSet r2 = lite.selectSQL("INSERT INTO stu(id, name) values(1, 'entao') ON CONFLICT(id) DO UPDATE SET name = 'entao' RETURNING id");
   println("upsert lastRowID: ", lite.lastInsertRowId);
   r2.dump();